requirements.txt
Flask==3.0.2
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
requests==2.31.0
